(TASK) Enable jruby pool write locker to borrow pool instances
This commit reworks the JRubyPool implementation to pull in all of the
concurrent constructs around a single ReentrantLock, the primary benefit
being to allow the holder of a write lock to be able to borrow instances
from the pool without running the risk of a deadlock due to competing
borrows that may be occurring from other threads.  The implementation
uses three condition variables to gate borrower and locker access to the
pool, similar to how LinkedBlockingDeque does for managing queue takers.

Where the underlying lock construct in the JRubyPool class would be a
ReentrantLock instead of a ReentrantReadWriteLock, this commit also
reworks a few of the unit tests to attempt borrows from non-locking
threads as an indication that the write lock is held, as opposed to
trying to interrogate the lock construct itself.