(TASK) Added lock tests and have clear() remove registered elements
This commit changes the implementation of `JRubyPool.clear` to remove
all registered elements associated with the elements being purged from
the queue.

This commit also adds several more tests to the `lockable-pool-test`
namespace, providing coverage for re-entrant calls to lock, borrowing
items with a timeout, inserting a pill, clearing the pool, and checking
the pool's remaining capacity.