(TASK) Use a separate ReentrantLock for writing
This commit introduces a second ReentrantLock to be used when a write
lock is requested.  The primary benefit of the change would be that a
would-be writer would be able to make its presence known to any
borrowers that happen to have been in the queue for the original
"shared" ReentrantLock ahead of the write lock.  Not sure that the
tradeoff of this potential improvement in how soon a writer gets
control is worth the additional complexity that would be needed to
manage a second lock around the use of the "shared" one.