(MAINT) Pin addressable gem to 2.4.x for Ruby 1.9.3 compatibility
Previously, bundle installs with the Gemfile for Ruby acceptance testing
failed when running under Ruby 1.9.3 because of an attempt to install
the public_suffix gem, which requires Ruby 2.  beaker 1.21.0 requires
google-api-client 0.7.1, which transitively requires addressable >=
2.3.2.  addressable, as of 2.5.0, started requiring the public_suffix
gem.

This commit pins addressable to ~> 2.4.0 in the Gemfile, allowing bundle
installs to succeed on Ruby 1.9.3.